sis-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1
geoalchemy2==0.14.2
shapely==2.0.2
geojson==3.1.0
python-dotenv==1.0.0
bcrypt==4.1.2
email-validator==2.1.0
slowapi==0.1.9
redis==5.0.1 